// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ADMIN: 'ADMIN',
  CREATOR: 'CREATOR',
  USER: 'USER'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const CourseStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED'
} as const

export type CourseStatus = (typeof CourseStatus)[keyof typeof CourseStatus]


export const LessonType = {
  VIDEO: 'VIDEO',
  TEXT: 'TEXT',
  QUIZ: 'QUIZ'
} as const

export type LessonType = (typeof LessonType)[keyof typeof LessonType]


export const MediaType = {
  VIDEO: 'VIDEO',
  IMAGE: 'IMAGE',
  PDF: 'PDF',
  AUDIO: 'AUDIO',
  TEXT: 'TEXT',
  OTHER: 'OTHER'
} as const

export type MediaType = (typeof MediaType)[keyof typeof MediaType]


export const PricingModel = {
  PAID: 'PAID',
  FREE: 'FREE',
  SUBSCRIPTION: 'SUBSCRIPTION'
} as const

export type PricingModel = (typeof PricingModel)[keyof typeof PricingModel]


export const OrderStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  FAILED: 'FAILED',
  REFUNDED: 'REFUNDED'
} as const

export type OrderStatus = (typeof OrderStatus)[keyof typeof OrderStatus]


export const EnrollmentSource = {
  PURCHASE: 'PURCHASE',
  SUBSCRIPTION: 'SUBSCRIPTION',
  FREE: 'FREE',
  ADMIN_GRANT: 'ADMIN_GRANT'
} as const

export type EnrollmentSource = (typeof EnrollmentSource)[keyof typeof EnrollmentSource]